model fixes wip

+++ libs/model/src/models/topic.ts
@@ -50,7 +50,7 @@ export default (
       weighted_voting: { type: Sequelize.STRING, allowNull: true },
       token_address: { type: Sequelize.STRING, allowNull: true },
       token_symbol: { type: Sequelize.STRING, allowNull: true },
-      vote_weight_multiplier: { type: Sequelize.INTEGER, allowNull: true },
+      vote_weight_multiplier: { type: Sequelize.FLOAT, allowNull: true },
     },
     {
       timestamps: true,
